requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.2.0
colorama==0.4.6
databases==0.9.0
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
pillow==11.2.1
pyasn1==0.4.8
pydantic==2.11.4
pydantic_core==2.33.2
PyMySQL==1.1.1
pyotp==2.9.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
qrcode==8.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
